GlobalVault/requirements.txt
requests==2.31.0
newspaper3k==0.2.8
python-dotenv==1.0.0
spacy==3.7.2
pandas==2.1.4
newspaper4k
